+++ Projetos/spring_mte/src/main/java/com/mte2023/spring_mte/services/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> findById(Long id){
+    public Optional<User> findById(Integer id){
         return userRepository.findById(id);
     }
 
@@ -33,12 +33,13 @@ public class UserService {
 
     public void saveAll(List<User> userList) { userRepository.saveAll(userList);}
     //delete methods
-    public void delete(Long id){
+    public void delete(Integer id){
         userRepository.deleteById(id);
     }
     //flush methods
     public void flush(){ userRepository.flush();}
 
+    /*
     public User update(Long id, User obj){
         User entity = userRepository.getReferenceById(id);
         updateData(entity, obj);
@@ -53,6 +54,7 @@ public class UserService {
 
     }
 
+     */
     //Logic Methods
     public void validate(User user) {
